Refactor: Export definition of the component
